requirements.txt
Flask
pandas
joblib
huggingface_hub
python-dotenv 